server/requirements.txt
fastapi[standard]>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.18
python-dotenv>=1.1.0

# LangChain ecosystem
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.3.0
langchain-ollama>=0.3.0
langchain-huggingface>=0.1.0
langchain-chroma>=0.2.0
langchain-community>=0.3.0

# Vector DB
chromadb>=0.6.0

# Embeddings
sentence-transformers>=3.4.0

# PDF parsing
pypdf>=5.0.0

# Word document parsing
python-docx>=1.1.0

# Settings management
pydantic-settings>=2.8.0

# Utilities
aiofiles>=24.1.0
